# doc/api/api.yaml
openapi: 3.0.3
info:
  description: ""
  version: "1.0.0"
  title: "Smart Doorbell API"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
servers:
  - url: "{protocol}://{host}/api/v1"
    variables:
      protocol:
        enum: [ https, http ]
        default: http
      host:
        default: localhost:8080
paths:
  /users:
    get:
      security:
        - OIDC: [ ]
      operationId: getAllUsers
      tags:
        - user
      responses:
        "200":
          description: "Successful operation"
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/UserResponse"
  /users/{userId}:
    parameters:
      - $ref: "#/components/parameters/UserId"
    get:
      security:
        - OIDC: [ ]
      operationId: getUserById
      tags:
        - user
      responses:
        "200":
          description: "Successful operation"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      security:
        - OIDC: [ ]
      operationId: deleteUser
      tags:
        - user
      responses:
        "204":
          description: "Successful deletion"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
    UserId:
      name: userId
      in: path
      description: "ID of user to which this doorbell belongs"
      required: true
      schema:
        type: integer
        format: int64
...
  responses:
    NotFound:
      description: "Entity was not found"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
  securitySchemes:
    OIDC:
      type: "openIdConnect"
      openIdConnectUrl: https://dev-gh2pzn93.us.auth0.com/.well-known/openid-configuration